[CWS] common status filling + platform specific
